… pkg/apis/storage/v1beta1/zz_generated.conversion.go
… pkg/apis/storage/validation/validation_test.go
… pkg/apis/storage/zz_generated.deepcopy.go
… pkg/registry/storage/csinode/strategy_test.go
… pkg/scheduler/BUILD
… pkg/scheduler/algorithm/predicates/csi_volume_predicate_test.go
… pkg/scheduler/algorithm/predicates/max_attachable_volume_predicate_test.go
… pkg/scheduler/factory/BUILD
… pkg/scheduler/factory/factory.go
… pkg/scheduler/factory/factory_test.go
… pkg/scheduler/internal/cache/cache.go
… pkg/scheduler/internal/cache/fake/fake_cache.go
… pkg/scheduler/internal/cache/interface.go
… pkg/scheduler/nodeinfo/BUILD
… pkg/scheduler/nodeinfo/node_info.go
… pkg/scheduler/scheduler.go
… pkg/scheduler/scheduler_test.go
… pkg/volume/csi/nodeinfomanager/BUILD
… pkg/volume/csi/nodeinfomanager/nodeinfomanager.go
… pkg/volume/csi/nodeinfomanager/nodeinfomanager_test.go
… plugin/pkg/auth/authorizer/rbac/bootstrappolicy/policy.go
… plugin/pkg/auth/authorizer/rbac/bootstrappolicy/testdata/cluster-roles.yaml
… staging/src/k8s.io/api/storage/v1beta1/generated.pb.go
… staging/src/k8s.io/api/storage/v1beta1/generated.proto
… staging/src/k8s.io/api/storage/v1beta1/types.go
… staging/src/k8s.io/api/storage/v1beta1/types_swagger_doc_generated.go
… staging/src/k8s.io/api/storage/v1beta1/zz_generated.deepcopy.go
… staging/src/k8s.io/api/testdata/HEAD/storage.k8s.io.v1beta1.CSINode.json
… staging/src/k8s.io/api/testdata/HEAD/storage.k8s.io.v1beta1.CSINode.pb
… staging/src/k8s.io/api/testdata/HEAD/storage.k8s.io.v1beta1.CSINode.yaml