@bertinatto 2019-05-10
> https://github.com/kubernetes/kubernetes/blob/master/pkg/registry/storage/csinode/strategy.go should also be modified to drop fields based on feature gates

Done