@bertinatto 2019-05-17
/test pull-kubernetes-verify
```
openapi-spec:

/home/prow/go/src/k8s.io/kubernetes/api/openapi-spec is out of date. Please run hack/update-openapi-spec.sh
```

The openapi-spec is updated and verify doesn't fail for me. I don't know why this test keeps failing.