@gnufied 2019-05-21
I think you do need to run `/hack/update-openapi-spec.sh` ? When I do that it does create a diff:


```
iff --git a/api/openapi-spec/swagger.json b/api/openapi-spec/swagger.json
index f1c5db1655..18261d0c21 100644
--- a/api/openapi-spec/swagger.json
+++ b/api/openapi-spec/swagger.json
@@ -18421,7 +18421,7 @@
   },
   "info": {
     "title": "Kubernetes",
-    "version": "v1.15.0"
+    "version": "v1.16.0"
   },
   "paths": {
     "/api/": {
```