@bertinatto 2019-05-25
/test pull-kubernetes-integration