@bertinatto 2019-05-26
/test pull-kubernetes-integration