@bertinatto 2019-05-29
/test pull-kubernetes-bazel-test