@liggitt 2019-05-31
fyi, ran this against the [in-progress API compatibility tests](https://github.com/kubernetes/kubernetes/pull/78309) and the changes in this PR do not affect round-tripping of 1.14.x-level data, so 👍