@liggitt 2019-06-03
FYI, we just added tests to verify we don't break compatibility with previously serialized data in #78309. Part of that keeps up-to-date serializations for current API types in `staging/src/k8s.io/api/testdata/HEAD`. That means that PRs that add fields (like this one), need to update those fixtures.

When this is rebased on master, run the following to update the CSINode fixtures:
```sh
UPDATE_COMPATIBILITY_FIXTURE_DATA=true go test ./vendor/k8s.io/api -run //HEAD
```