@msau42 2019-06-04
I ran the secrets and secretsWithCSINode benchmarks with and without this pr on my local desktop with 12 cores.  And saw no signficant regression (within +/- 3%).  I also ran it with and without optimizations, and saw no significant difference either.