@jsafrane 2019-06-05
I tested with today's version (morning) and numbers seem to be better:


Test | Before | After | Diff
-- | -- | -- | --
BenchmarkSchedulingCSIPVs | 61923461.6 | 69684354 | 12.53%
BenchmarkSchedulingInTreePVs | 76875485.9 | 78601283.5 | 2.24%
BenchmarkSchedulingMigratedInTreePVs | 77419368.5 | 79149593.9 | 2.23%
BenchmarkSchedulingPodAntiAffinity | 36619170.1 | 36661907.3 | 0.12%
BenchmarkSchedulingSecrets | 29228343.3 | 28981172.3 | -0.85%
BenchmarkSchedulingSecretsWithCSINodes | 28915431.4 | 30266347 | 4.67%