@jsafrane 2019-06-19
Fresh numbers from today's morning:

| Test | Before (ns/op) | After (ns/op) | Increase
-- | -- | -- | --
BenchmarkSchedulingCSIPVs | 58977913.6 | 64527353.9 | 9.41%
BenchmarkSchedulingInTreePVs | 74486511.8 | 75411026.2 | 1.24%
BenchmarkSchedulingMigratedInTreePVs | 74846972.9 | 74396914.4 | -0.60%
BenchmarkSchedulingPodAntiAffinity | 36280202.1 | 36486112.5 | 0.57%
BenchmarkSchedulingSecrets | 29108418 | 29609940.5 | 1.72%
BenchmarkSchedulingSecretsWithCSINodes | 29691660.5 | 30196475 | 1.70%

Only scheduling of pods with CSI volumes is ~9% slower. The rest (incl. pod with Secrets and in-tree volumes) is negligible.