@jsafrane 2019-06-24
Just a note to the benchmark numbers: they were taken on 4 core machine. With 8 cores, there is no performance regression. Scheduler seems to run (some?) predicates in parallel and the updated CSIMaxVolumeLimitChecker predicate is still faster than the slowest already existing predicate.