@ravisantoshgudimetla 2019-06-24
> With 8 cores, there is no performance regression. Scheduler seems to run (some?) predicates 

Yes, scheduler runs predicates in parallel. You mean on 4 core machine, you noticed performance regression of 9% whereas on 8 core there is no performance regression. That sounds reasonable. LGTM from my side

/lgtm

@bsalamat - Do you mind doing a final review since you were involved in the design?