@tedyu 2019-06-26
@jsafrane @msau42 
what do you think of the proposal of caching csiVolumeLimits inside NodeInfo ?

thanks