@msau42 2019-06-26
@tedyu the number of different drivers installed on a node are not expected to be very high. I'm not sure if we will see much performance benefit from this.  You can try running the benchmarks that @jsafrane developed to see.  https://github.com/kubernetes/kubernetes/pull/77397